% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MorphoTreat_PACKAGE.R
\docType{package}
\name{MorphoTreat}
\alias{MorphoTreat}
\title{MorphoTreat}
\description{
An R package for Identifying Antimalarial Drug Treatments Based on the Blood Stage Morphology of P. falciparum
}
