% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShowPredict.R
\name{ShowPredict}
\alias{ShowPredict}
\title{ShowPredict}
\usage{
ShowPredict(img.name, output.path)
}
\arguments{
\item{img.name}{the input image file name}

\item{output.path}{the YOLO output path}
}
\description{
ShowPredict shows the image of the predict result.
}
